requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
tiktoken>=0.5.0
redis>=5.0.0
pymysql>=1.1.0
cryptography>=41.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-httpx>=0.26.0
aiofiles>=23.2.0
